refactor: move types outside src
