+++ src/components/personal/students/plannings/TrainingPlanningCard/TrainingPlanningCard.tsx
@@ -1,4 +1,4 @@
-import { TrainingPlanningWithDetails } from "../../../../../types/schema"
+import { TrainingPlanningWithDetails } from "../../../../../../types/schema"
 import styles from './TrainingPlanningCard.module.scss';
 import badgeStyles from '../../../../../styles/components/badge.module.scss';
 import { getTrainingPlanningMuscleGroupsPerTraining } from "../../../../../lib/training";
