+++ src/lib/swr/usePersonalStudents.ts
@@ -1,6 +1,6 @@
 import axios from "axios";
 import useSWR, { useSWRConfig } from "swr";
-import { PersonalStudentWithTrainings } from "../../types/schema";
+import { PersonalStudentWithTrainings } from "../../../types/schema";
 
 const fetchPersonalStudents = (email: string) => (url: string) => axios.get<PersonalStudentWithTrainings[]>(url, {
   params: {
