+++ src/pages/fill-profile/index.tsx
@@ -4,7 +4,7 @@ import Head from 'next/head'
 import { useRouter } from 'next/router'
 import React, { FormEventHandler, useState } from 'react'
 import { MdSend } from 'react-icons/md'
-import { NextPageWithAuth } from '../../types/page'
+import { NextPageWithAuth } from '../../../types/page'
 import Header from '../../components/Header'
 import Layout from '../../components/Layout'
 import styles from '../../styles/pages/fill-data.module.scss'
